# Use malachite for modular inverse