--- a/crypto/muhash/src/u3072.rs
+++ b/crypto/muhash/src/u3072.rs
@@ -147,8 +147,13 @@ impl U3072 {
             a.full_reduce();
         }
         // The only value that doesn't have a multiplicative inverse is 0, and 0/x is 0.
-        let inv = Uint3072(a.limbs).mod_inverse(Self::UINT_PRIME).unwrap_or_default();
-        Self { limbs: inv.0 }
+        let inv = Self { limbs: Uint3072(a.limbs).mod_inverse(Self::UINT_PRIME).unwrap_or_default().0 };
+        if cfg!(debug_assertions) {
+            let mut one = inv;
+            one *= a;
+            assert_eq!(one, Self::one());
+        }
+        inv
     }
 
     fn div(&mut self, other: &Self) {