--- a/math/fuzz/fuzz_targets/u256.rs
+++ b/math/fuzz/fuzz_targets/u256.rs
@@ -135,7 +135,7 @@ fuzz_target!(|data: &[u8]| {
         let ((lib1, native1), (lib2, native2)) = loop {
             let (lib1, native1) = try_opt!(generate_ints(&mut data));
             let (lib2, native2) = try_opt!(generate_ints(&mut data));
-            if lib1 != 1u64 || lib2 != 1u64 {
+            if lib1 < lib2 && lib1 != 0u64 {
                 break ((lib1, native1), (lib2, native2));
             }
         };