--- a/math/src/uint.rs
+++ b/math/src/uint.rs
@@ -352,15 +352,15 @@ macro_rules! construct_uint {
 
             /// Converts a Self::BYTES*2 hex string interperted as big endian, into a Uint
             #[inline]
-            pub fn from_hex(hex: &str) -> Result<Self, faster_hex::Error> {
+            pub fn from_hex(hex: &str) -> Result<Self, $crate::uint::faster_hex::Error> {
                 if hex.len() > Self::BYTES * 2 {
-                    return Err(faster_hex::Error::InvalidLength(hex.len()));
+                    return Err($crate::uint::faster_hex::Error::InvalidLength(hex.len()));
                 }
                 let mut out = [0u8; Self::BYTES];
                 let mut input = [b'0'; Self::BYTES * 2];
                 let start = input.len() - hex.len();
                 input[start..].copy_from_slice(hex.as_bytes());
-                faster_hex::hex_decode(&input, &mut out)?;
+                $crate::uint::faster_hex::hex_decode(&input, &mut out)?;
                 Ok(Self::from_be_bytes(out))
             }
         }
@@ -774,7 +774,7 @@ macro_rules! construct_uint {
 
             #[inline]
             fn deserialize_in_place<D: $crate::uint::serde::Deserializer<'de>>(
-                deserializer: D,
+                    deserializer: D,
                 place: &mut Self,
             ) -> Result<(), D::Error> {
                 use core::fmt;
@@ -942,6 +942,50 @@ mod tests {
                 str_buf.write_fmt(format_args!("{mine:032x}")).unwrap();
                 assert_eq!(mine, Uint128::from_hex(&str_buf).unwrap());
             }
+        }
+    }
+
+    #[test]
+    fn test_mod_inv() {
+        use core::cmp::Ordering;
+        let mut rng = ChaCha8Rng::from_seed([0; 32]);
+        let mut buf = [0u8; 16];
+        for _ in 0..50_000 {
+            rng.fill_bytes(&mut buf);
+            let uint1 = Uint128::from_le_bytes(buf);
+            rng.fill_bytes(&mut buf);
+            let uint2 = Uint128::from_le_bytes(buf);
+            let (bigger, smaller) = match uint1.cmp(&uint2) {
+                Ordering::Greater => (uint1, uint2),
+                Ordering::Less => (uint2, uint1),
+                Ordering::Equal => continue,
+            };
+            let inv = smaller.mod_inverse(bigger);
+            if let Some(inv) = inv {
+                assert_eq!(prod_bin(inv, smaller, bigger), 1u64);
+            }
+        }
+
+        fn sum(x: Uint128, y: Uint128, m: Uint128) -> Uint128 {
+            let res = x.overflowing_add(y).0;
+            if res < x || res >= m {
+                res.overflowing_sub(m).0
+            } else {
+                res
+            }
+        }
+        fn prod_bin(x: Uint128, y: Uint128, m: Uint128) -> Uint128 {
+            if y == 1u64 {
+                return x;
+            } else if y == 0u64 {
+                return Uint128::ZERO;
+            }
+            let mut res = prod_bin(x, y >> 1, m);
+            res = sum(res, res, m);
+            if (y.as_u64() & 1) == 1 {
+                res = sum(res, x, m);
+            }
+            res
         }
     }
 }