This uses the new https://github.com/mhogrefe/malachite crate for computing modular inverses, in the future, we'd want to replace that with our own implementation of Lehmer's algorithm.
Benchmark before on my machine with turbo-boost off:
```
Uint3072/mod_inv Muhash prime
                        time:   [254.88 ms 257.58 ms 260.49 ms]
                        change: [+0.5522% +1.8513% +3.4027%] (p = 0.01 < 0.05)
                        Change within noise threshold.
```

After: 

```
Uint3072/mod_inv Muhash prime
                        time:   [1.5437 ms 1.5536 ms 1.5654 ms]
                        change: [-99.404% -99.396% -99.388%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
```

I will let the fuzzers run throughout the weekend

(Thanks @michaelsutton For pointing this library out and suggesting we'll use it here)